complete transform custom controls
Completed a basic transformControls on the test cube and put the controls in a separate file from the object
